requirements.txt
streamlit
opencv-python-headless
numpy
scikit-learn
Pillow
matplotlib
opencv-python-headless 
opencv-python
